Enable Consult to be operated from external package, and test that predicate/rule can be consulted
Without operators the Consult function can't be called and correctly completed from an external package. The existing tests add the operators manually before calling consult.
But the tests in text_test.go rely on the fact they are in the "engine" package. The SetPrologOperators is a - potentially expandable - set of default operators needed to Consult prolog files.

The rawtext map added to the VM structure is a placeholder for the next pull request.