Enable a file and added predicates to be saved to a prolog file
An original prolog file that is consulted may be available. However if a file is appended to, or created dynamically the prolog file isn't easily exported.
The functions added to builtin.go and text.go allow this. This may enable the results from the ichiban/prolog go prolog interpreter and other interpreters to be compared.